+++ changelog.md
@@ -2,6 +2,7 @@
 --
 * Add the 'Attachment' model that makes it easier to link your own models to
   assets from the media module.
+* Make it easier to download the original file in the asset module.
 
 
 7.5.1 - 2018-07-26